Move expression flags to u64s (60 maximum)
In this commit I attempted to overcome the limit of only being able to
access 31 out of Miitomo's 70 expression flags by making expressionFlag
a u64. It didn't take much effort and, this still isn't a real solution
due to it only getting about 60 of them in, but I'm pushing it anyway.
Expect me to revert this, or something.

+++ include/nn/ffl/FFLiMaskTextures.h
@@ -24,11 +24,11 @@ class   FFLiResourceLoader;
 class   FFLiResourceManager;
 class   FFLiShaderCallback;
 
-FFLExpression FFLiInitMaskTextures(FFLiMaskTextures* pMaskTextures, u32 expressionFlag, u32 resolution, bool enableMipMap);
+FFLExpression FFLiInitMaskTextures(FFLiMaskTextures* pMaskTextures, u64 expressionFlag, u32 resolution, bool enableMipMap);
 void FFLiDeleteMaskTextures(FFLiMaskTextures* pMaskTextures);
 
-FFLResult FFLiInitTempObjectMaskTextures(FFLiMaskTexturesTempObject* pObject, const FFLiMaskTextures* pMaskTextures, const FFLiCharInfo* pCharInfo, u32 expressionFlag, u32 resolution, bool enableMipMap, FFLiResourceLoader* pResLoader);
-void FFLiDeleteTempObjectMaskTextures(FFLiMaskTexturesTempObject* pObject, u32 expressionFlag, FFLResourceType resourceType);
+FFLResult FFLiInitTempObjectMaskTextures(FFLiMaskTexturesTempObject* pObject, const FFLiMaskTextures* pMaskTextures, const FFLiCharInfo* pCharInfo, u64 expressionFlag, u32 resolution, bool enableMipMap, FFLiResourceLoader* pResLoader);
+void FFLiDeleteTempObjectMaskTextures(FFLiMaskTexturesTempObject* pObject, u64 expressionFlag, FFLResourceType resourceType);
 
 void FFLiRenderMaskTextures(FFLiMaskTextures* pMaskTextures, FFLiMaskTexturesTempObject* pObject, const FFLiShaderCallback* pCallback
 #if RIO_IS_CAFE
